Module oci_distribution::annotations

OCI Annotation key constants, taken from: https://github.com/opencontainers/image-spec/blob/master/annotations.md#pre-defined-annotation-keys

Constants

ORG_OPENCONTAINERS_IMAGE_AUTHORS

Contact details of the people or organization responsible for the image (freeform string)

ORG_OPENCONTAINERS_IMAGE_BASE_DIGEST

Digest of the image this image is based on (string)

ORG_OPENCONTAINERS_IMAGE_BASE_NAME

If the image.base.name annotation is specified, the image.base.digest annotation SHOULD be the digest of the manifest referenced by the image.ref.name annotation.

ORG_OPENCONTAINERS_IMAGE_CREATED

Date and time on which the image was built (string, date-time as defined by RFC 3339)

ORG_OPENCONTAINERS_IMAGE_DESCRIPTION

Human-readable description of the software packaged in the image (string)

ORG_OPENCONTAINERS_IMAGE_DOCUMENTATION

URL to get documentation on the image (string)

ORG_OPENCONTAINERS_IMAGE_LICENSES

License(s) under which contained software is distributed as an SPDX License Expression

ORG_OPENCONTAINERS_IMAGE_REF_NAME

Name of the reference for a target (string)

ORG_OPENCONTAINERS_IMAGE_REVISION

Source control revision identifier for the packaged software

ORG_OPENCONTAINERS_IMAGE_SOURCE

URL to get source code for building the image (string)

ORG_OPENCONTAINERS_IMAGE_TITLE

Human-readable title of the image (string)

ORG_OPENCONTAINERS_IMAGE_URL

URL to find more information on the image (string)

ORG_OPENCONTAINERS_IMAGE_VENDOR

Name of the distributing entity, organization or individual

ORG_OPENCONTAINERS_IMAGE_VERSION

Version of the packaged software