Module oci_distribution::client

OCI distribution client

Note: This client is very feature poor. We hope to expand this to be a complete OCI distribution client in the future.

Structs

Certificate

A x509 certificate

Client

The OCI client connects to an OCI registry and fetches OCI images.

ClientConfig

A client configuration

ImageData

The data for an image or module.

ImageLayer

The data and media type for an image layer

PushResponse

The data returned by an OCI registry after a successful push operation is completed

Enums

CertificateEncoding

The encoding of the certificate

ClientProtocol

The protocol that the client should use to connect

Traits

ClientConfigSource

A source that can provide a ClientConfig. If you are using this crate in your own application, you can implement this trait on your configuration type so that it can be passed to Client::from_source.

Functions

current_platform_resolver

A platform resolver that chooses the first variant matching the running OS/Arch, if present. Doesn’t currently handle platform.variants.

linux_amd64_resolver

A platform resolver that chooses the first linux/amd64 variant, if present