Module oci_distribution::manifest

OCI Manifest

Structs

ImageIndexEntry

The manifest entry of an ImageIndex.

OciDescriptor

The OCI descriptor is a generic object used to describe other objects.

OciImageIndex

The image index is a higher-level manifest which points to specific image manifest.

OciImageManifest

The OCI image manifest describes an OCI image.

Platform

Platform specific fields of an Image Index manifest entry.

Versioned

Versioned provides a struct with the manifest’s schemaVersion and mediaType. Incoming content with unknown schema versions can be decoded against this struct to check the version.

Enums

OciManifest

An image, or image index, OCI manifest

Constants

IMAGE_CONFIG_MEDIA_TYPE

The mediatype for an image config (manifest).

IMAGE_DOCKER_CONFIG_MEDIA_TYPE

The mediatype that Docker uses for image configs.

IMAGE_DOCKER_LAYER_GZIP_MEDIA_TYPE

The mediatype that Docker uses for a layer that is gzipped.

IMAGE_LAYER_GZIP_MEDIA_TYPE

The mediatype for a layer that is gzipped.

IMAGE_LAYER_MEDIA_TYPE

The mediatype for a layer.

IMAGE_LAYER_NONDISTRIBUTABLE_GZIP_MEDIA_TYPE

The mediatype for a layer that is nondistributable and gzipped.

IMAGE_LAYER_NONDISTRIBUTABLE_MEDIA_TYPE

The mediatype for a layer that is nondistributable.

IMAGE_MANIFEST_LIST_MEDIA_TYPE

The mediatype for an docker v2 shema 2 manifest list.

IMAGE_MANIFEST_MEDIA_TYPE

The mediatype for an docker v2 schema 2 manifest.

OCI_IMAGE_INDEX_MEDIA_TYPE

The mediatype for an OCI image index manifest.

OCI_IMAGE_MEDIA_TYPE

The mediatype for an OCI image manifest.

WASM_CONFIG_MEDIA_TYPE

The mediatype for a WASM image config.

WASM_LAYER_MEDIA_TYPE

The mediatype for WASM layers.